requirements_qwen.txt
absl-py==2.1.0
accelerate==1.3.0
annotated-types==0.7.0
certifi==2024.12.14
charset-normalizer==3.4.1
colour==0.1.5
contourpy==1.3.1
cycler==0.12.1
deepspeed==0.16.3
dtreeviz==2.2.2
einops==0.8.0
exceptiongroup==1.2.2
filelock==3.17.0
fonttools==4.55.6
fsspec==2024.12.0
graphviz==0.20.3
grpcio==1.70.0
hjson==3.1.0
huggingface-hub==0.27.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
joblib @ 
kiwisolver==1.4.8
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mkl-service==2.4.0
mkl_fft @ 
mkl_random @ 
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
ninja==1.11.1.3
numpy @ 
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.2.3
peft==0.14.0
pillow==11.1.0
pluggy==1.5.0
protobuf==5.29.3
psutil==6.1.1
py-cpuinfo==9.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn @ 
scipy @ 
seaborn==0.13.2
six==1.17.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl @ 
tiktoken==0.8.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.48.1
transformers-stream-generator==0.0.4
triton==3.1.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
